feat: add DCT32x16 and DCT16x32 transform implementation
Adds full implementation for DCT32x16 and DCT16x32 transforms:
- Forward DCT functions (dct_32x16, dct_16x32)
- DC extraction (dc_from_dct_32x16, dc_from_dct_16x32)
- Quantization weights with 8-band parametric generation
- Strategy codes (10, 11) and coverage arrays
- Block context map entries for proper decoder context
- Coefficient ordering (512 coefficients with 4x2/2x4 LLF region)
- IDCT fallback for pixel-domain loss estimation
- AC group tokenization support

Strategy selection currently disabled pending investigation of
interaction with LZ77 backref encoding (decoder errors with
InvalidAnsStream at d>=3.0 with LZ77 enabled).

All tests pass with selection disabled. The infrastructure is ready
for enabling once the LZ77 interaction issue is resolved.

+++ jxl_enc/src/tiny/lz77.rs
@@ -1065,7 +1065,7 @@ mod tests {
         }
 
         let _result_no_mult = apply_lz77_backref(&tokens, 1, false, 0);
-        let result_with_mult = apply_lz77_backref(&tokens, 1, false, image_width as i32);
+        let result_with_mult = apply_lz77_backref(&tokens, 1, false, image_width);
 
         // Both should find matches; with multiplier might be more efficient
         // but the main test is that it doesn't crash and produces valid output
